+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- [2025-11-12] [Support default, nullable, and optional in shapeOf](https://github.com/rubriclab/shapes/commit/8b755019e28b7bc68587a8d553d93ffff3b6e747)
 - [2025-09-19] [Return-type stringify function](https://github.com/rubriclab/shapes/commit/e1533f6af045115b549a9df8569ac1fd702e9a4a)
 - [2025-09-17] [bump](https://github.com/RubricLab/shapes/commit/f65e4d6300f3cb7e67feb7fd979e0355f2170184)
 - [2025-07-25] [rework scoped shape](https://github.com/RubricLab/shapes/commit/76f70ee65df6547626cc37b03713b4e0d8b65d23)